pipeline/dagster_pipeline/requirements.txt
# Dagster Pipeline Requirements
dagster>=1.6.0
dagster-webserver>=1.6.0
dagster-snowflake>=0.22.0
dagster-aws>=0.22.0
dagster-dbt>=0.22.0

# Data Processing
polars>=0.20.0
pandas>=2.0.3
numpy>=1.24.3
pyarrow>=12.0.0

# ML
xgboost==1.7.6
scikit-learn>=1.3.2

# Snowflake
snowflake-connector-python[pandas]>=3.6.0

# AWS
boto3>=1.34.0

# HTTP
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
